[#3514] improvement(flink-connector): add flink-connector-runtime to build flink connector (#4791)
### What changes were proposed in this pull request?

- add flink runtime connector

### Why are the changes needed?

Fix: #3514 

### Does this PR introduce _any_ user-facing change?

- no

### How was this patch tested?

- local test

---------
