+++ flink-connector/flink/src/test/resources/flink-tests/hive-site.xml
